feat : init global exception handler
