--- src/main/java/com/example/demo/auth/AuthController.java
@@ -1,32 +1,0 @@
-package com.example.demo.auth;
-
-import com.example.demo.auth.model.AuthResponse;
-import com.example.demo.auth.model.LoginRequest;
-import com.example.demo.auth.model.RegisterRequest;
-import com.example.demo.auth.model.RegisterResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/v1/auth")
-@RequiredArgsConstructor
-public class AuthController {
-    private final AuthService authService;
-
-    @PostMapping("/register")
-    public ResponseEntity<RegisterResponse> register(@RequestBody RegisterRequest request) {
-        RegisterResponse registerResponse = authService.register(request);
-        return new ResponseEntity<>(registerResponse, HttpStatus.OK);
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<?> login(@RequestBody LoginRequest request) {
-        AuthResponse authenticationResponse = authService.login(request);
-        return new ResponseEntity<>(authenticationResponse, HttpStatus.OK);
-    }
-}
