--- src/main/java/com/example/demo/auth/model/AuthResponse.java
@@ -1,15 +1,0 @@
-package com.example.demo.auth.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthResponse {
-    private String token;
-    private String message;
-}
