--- src/main/java/com/example/demo/auth/model/RegisterRequest.java
@@ -1,16 +1,0 @@
-package com.example.demo.auth.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class RegisterRequest {
-    private String email;
-    private String password;
-    private String role;
-}
