--- src/main/java/com/example/demo/auth/model/RegisterResponse.java
@@ -1,15 +1,0 @@
-package com.example.demo.auth.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class RegisterResponse {
-    private String message;
-    private String userId;
-}
