--- src/main/java/com/example/demo/auth_user/AuthUserMapper.java
@@ -1,9 +1,0 @@
-package com.example.demo.auth_user;
-
-import com.example.demo.auth_user.dto.AuthUserDTO;
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel = "spring")
-public interface AuthUserMapper {
-    AuthUserDTO mapFromEntity(AuthUser authUser);
-}
