--- src/main/java/com/example/demo/auth_user/AuthUserRepository.java
@@ -1,13 +1,0 @@
-package com.example.demo.auth_user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface AuthUserRepository extends JpaRepository<AuthUser, Long> {
-    Optional<AuthUser> findByEmail(String email);
-    Optional<AuthUser> findById(UUID id);
-}
