--- src/main/java/com/example/demo/auth_user/dto/AuthUserDTO.java
@@ -1,10 +1,0 @@
-package com.example.demo.auth_user.dto;
-
-import java.util.UUID;
-
-public record AuthUserDTO(
-        UUID id,
-        String email,
-        String role
-) {
-}
