--- src/main/java/com/example/demo/core/enums/RoleEnum.java
@@ -1,6 +1,0 @@
-package com.example.demo.core.enums;
-
-public enum RoleEnum {
-    ADMIN,
-    USER,
-}
