--- src/main/java/com/example/demo/security/config/ApplicationConfig.java
@@ -1,21 +1,0 @@
-package com.example.demo.security.config;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@RequiredArgsConstructor
-@ComponentScan
-public class ApplicationConfig implements WebMvcConfigurer {
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-                .allowedOriginPatterns("*")
-                .allowedMethods("HEAD", "GET", "POST", "PUT", "DELETE", "PATCH")
-                .allowedHeaders("Authorization", "Cache-Control", "Content-Type")
-                .allowCredentials(true);
-    }
-}
