+++ src/main/java/com/jeyofdev/yellow_berry/YellowBerryApplication.java
@@ -1,0 +1,25 @@
+package com.jeyofdev.yellow_berry;
+
+import com.jeyofdev.yellow_berry.config.DatabaseConfig;
+import com.jeyofdev.yellow_berry.util.DatabaseInitializer;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ApplicationContext;
+
+@SpringBootApplication
+public class YellowBerryApplication {
+
+	public static void main(String[] args) {
+		ApplicationContext context = SpringApplication.run(YellowBerryApplication.class, args);
+		DatabaseConfig dbConfig = context.getBean(DatabaseConfig.class);
+
+		// Check if database exist
+		DatabaseInitializer.initializeDatabase(
+				"jdbc:postgresql://localhost:5432/postgres",
+				dbConfig.getDbUser(),
+				dbConfig.getDbPassword(),
+				dbConfig.getDbName()
+		);
+	}
+
+}
