+++ src/main/java/com/jeyofdev/yellow_berry/auth/model/AuthResponse.java
@@ -1,0 +1,15 @@
+package com.jeyofdev.yellow_berry.auth.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthResponse {
+    private String token;
+    private String message;
+}
