+++ src/main/java/com/jeyofdev/yellow_berry/auth/model/RegisterRequest.java
@@ -1,0 +1,16 @@
+package com.jeyofdev.yellow_berry.auth.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class RegisterRequest {
+    private String email;
+    private String password;
+    private String role;
+}
