+++ src/main/java/com/jeyofdev/yellow_berry/auth_user/AuthUser.java
@@ -1,0 +1,47 @@
+package com.jeyofdev.yellow_berry.auth_user;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import jakarta.persistence.*;
+import lombok.*;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+import java.util.List;
+import java.util.UUID;
+
+@Entity
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@ToString
+@Table(name = "\"user\"")
+public class AuthUser implements UserDetails {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private UUID id;
+
+    @Column(name = "email", columnDefinition = "VARCHAR(100)")
+    private String email;
+
+    @JsonIgnore
+    @Column(name = "password", columnDefinition = "VARCHAR(255)")
+    private String password;
+
+    @JsonIgnore
+    @Column(name = "role", columnDefinition = "VARCHAR(20)")
+    private String role;
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return List.of(new SimpleGrantedAuthority(role));
+    }
+
+    @Override
+    public String getUsername() {
+        return email;
+    }
+}
