+++ src/main/java/com/jeyofdev/yellow_berry/auth_user/AuthUserRepository.java
@@ -1,0 +1,13 @@
+package com.jeyofdev.yellow_berry.auth_user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+@Repository
+public interface AuthUserRepository extends JpaRepository<AuthUser, Long> {
+    Optional<AuthUser> findByEmail(String email);
+    Optional<AuthUser> findById(UUID id);
+}
