+++ src/main/java/com/jeyofdev/yellow_berry/auth_user/dto/AuthUserDTO.java
@@ -1,0 +1,10 @@
+package com.jeyofdev.yellow_berry.auth_user.dto;
+
+import java.util.UUID;
+
+public record AuthUserDTO(
+        UUID id,
+        String email,
+        String role
+) {
+}
