+++ src/main/java/com/jeyofdev/yellow_berry/core/enums/RoleEnum.java
@@ -1,0 +1,6 @@
+package com.jeyofdev.yellow_berry.core.enums;
+
+public enum RoleEnum {
+    ADMIN,
+    USER,
+}
