+++ src/main/java/com/jeyofdev/yellow_berry/exception/NotFoundException.java
@@ -1,0 +1,15 @@
+package com.jeyofdev.yellow_berry.exception;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+
+    public NotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public NotFoundException(Throwable cause) {
+        super(cause);
+    }
+}
