+++ src/main/java/com/jeyofdev/yellow_berry/exception/UsernameAlreadyTakenException.java
@@ -1,0 +1,15 @@
+package com.jeyofdev.yellow_berry.exception;
+
+public class UsernameAlreadyTakenException extends RuntimeException {
+    public UsernameAlreadyTakenException(String message) {
+        super(message);
+    }
+
+    public UsernameAlreadyTakenException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public UsernameAlreadyTakenException(Throwable cause) {
+        super(cause);
+    }
+}
