+++ src/main/java/com/jeyofdev/yellow_berry/util/DatabaseInitializer.java
@@ -1,0 +1,27 @@
+package com.jeyofdev.yellow_berry.util;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+import java.sql.Statement;
+import java.text.MessageFormat;
+
+public class DatabaseInitializer {
+    public static void initializeDatabase(String jdbcUrl, String user, String password, String dbName) {
+        String createDbQuery = MessageFormat.format("CREATE DATABASE {0}", dbName);
+        try (Connection connection = DriverManager.getConnection(jdbcUrl, user, password);
+             Statement statement = connection.createStatement()) {
+
+            // If the database does not exist, create it
+            statement.executeUpdate(createDbQuery);
+            System.out.println("Database created successfully!");
+        } catch (SQLException e) {
+            if (e.getSQLState().equals("42P04")) { // Code error for "Database already exists"
+                System.out.println("The database already exists.");
+            } else {
+                System.err.println(MessageFormat.format("Error creating database : {0}", e.getMessage()));
+                throw new RuntimeException(e);
+            }
+        }
+    }
+}
